fix: update batch region to address quota alerts
