+++ batch/snippets/src/main/java/com/example/batch/DeleteJob.java
@@ -12,49 +12,47 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// [START batch_get_task]
+package com.example.batch;
 
+// [START batch_delete_job]
 import com.google.cloud.batch.v1.BatchServiceClient;
-import com.google.cloud.batch.v1.Task;
-import com.google.cloud.batch.v1.TaskName;
 import java.io.IOException;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
-public class GetTask {
+public class DeleteJob {
 
-  public static void main(String[] args) throws IOException {
+  public static void main(String[] args)
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     // TODO(developer): Replace these variables before running the sample.
     // Project ID or project number of the Cloud project you want to use.
     String projectId = "YOUR_PROJECT_ID";
+
     // Name of the region hosts the job.
     String region = "europe-central2";
-    // The name of the job you want to retrieve information about.
-    String jobName = "JOB_NAME";
-    // The name of the group that owns the task you want to check. Usually it's `group0`.
-    String groupName = "group0";
-    // Number of the task you want to look up.
-    int taskNumber = 0;
 
-    getTask(projectId, region, jobName, groupName, taskNumber);
+    // The name of the job that you want to delete.
+    String jobName = "JOB_NAME";
+
+    deleteJob(projectId, region, jobName);
   }
 
-  // Retrieve information about a Task.
-  public static void getTask(String projectId, String region, String jobName, String groupName,
-      int taskNumber) throws IOException {
+  // Triggers the deletion of a Job.
+  public static void deleteJob(String projectId, String region, String jobName)
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     // Initialize client that will be used to send requests. This client only needs to be created
     // once, and can be reused for multiple requests. After completing all of your requests, call
     // the `batchServiceClient.close()` method on the client to safely
     // clean up any remaining background resources.
     try (BatchServiceClient batchServiceClient = BatchServiceClient.create()) {
 
-      Task task = batchServiceClient.getTask(TaskName.newBuilder()
-          .setProject(projectId)
-          .setLocation(region)
-          .setJob(jobName)
-          .setTaskGroup(groupName)
-          .setTask(String.valueOf(taskNumber))
-          .build());
-      System.out.printf("Retrieved task information: %s", task.getName());
+      // Construct the parent path of the job.
+      String name = String.format("projects/%s/locations/%s/jobs/%s", projectId, region, jobName);
+
+      batchServiceClient.deleteJobAsync(name).get(5, TimeUnit.MINUTES);
+      System.out.printf("Delete the job: %s", jobName);
     }
   }
 }
-// [END batch_get_task]
+// [END batch_delete_job]
