+++ batch/snippets/src/test/java/com/example/batch/Util.java
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+package com.example.batch;
+
 import static com.google.common.truth.Truth.assertThat;
 
 import com.google.cloud.batch.v1.BatchServiceClient;
